add mipmap library and correctly handle different processors
