+++ CMakeLists.txt
@@ -1071,8 +1071,7 @@ IF (OPENGL_FOUND)
     src/texcmp/squish/singlecolourfitfast.cpp
     src/texcmp/squish/twocolourfitfast.cpp
     src/texcmp/squish/squish.cpp
-    src/texcmp/etcpak.cpp
-    src/texcmp/mipmap.cpp)
+    src/texcmp/etcpak.cpp)
 
   INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/src/texcmp/lz4)
   INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/src/texcmp/squish)
@@ -1086,11 +1085,7 @@ IF (OPENGL_FOUND)
       IF (( ARCH MATCHES "i386" OR ARCH MATCHES "amd64" OR ARCH MATCHES "x86_64") AND NOT QT_ANDROID)
         # using sse which makes the compression run about 50% faster
         message (STATUS "Compiling texture compression library with sse support")
-#        set_property(TARGET TEXCMP PROPERTY COMPILE_FLAGS "-O3 -msse2 -DSQUISH_USE_SSE=2")
-#        set_property(TARGET TEXCMP PROPERTY COMPILE_FLAGS "-O3 -msse2 -mssse3 -DSQUISH_USE_SSE=2")
         set_property(TARGET TEXCMP PROPERTY COMPILE_FLAGS "-O3 -msse2 -DSQUISH_USE_SSE=2")
-
-        set_source_files_properties(src/texcmp/mipmap.cpp PROPERTIES COMPILE_FLAGS "")
       ELSE ()
         # use standard optimizations for other architectures
         set_property(TARGET TEXCMP PROPERTY COMPILE_FLAGS "-O3")
@@ -1101,6 +1096,39 @@ IF (OPENGL_FOUND)
         # using sse which makes the compression run about 50% faster
         message (STATUS "Compiling texture compression library with sse support")
         set_property(TARGET TEXCMP PROPERTY COMPILE_FLAGS "/arch:SSE2 -DSQUISH_USE_SSE=2")
+      ENDIF ()
+  ENDIF (NOT MSVC)
+
+
+  SET(SRC_MIPMAP
+    src/mipmap/mipmap.c
+    src/mipmap/mipmap_sse.c
+    src/mipmap/mipmap_sse2.c
+    src/mipmap/mipmap_ssse3.c
+    src/mipmap/mipmap_avx2.c)
+
+  ADD_LIBRARY(MIPMAP ${SRC_MIPMAP})
+  SET(EXTRA_LIBS ${EXTRA_LIBS} MIPMAP)
+
+  # always use optimizations for the texture compression which is much faster
+  IF ( NOT MSVC )
+      set_property(TARGET MIPMAP PROPERTY COMPILE_FLAGS "-O3")
+
+      IF (( ARCH MATCHES "i386" OR ARCH MATCHES "amd64" OR ARCH MATCHES "x86_64") AND NOT QT_ANDROID)
+        set_source_files_properties(src/mipmap/mipmap_sse.c PROPERTIES COMPILE_FLAGS "-msse")
+        set_source_files_properties(src/mipmap/mipmap_sse2.c PROPERTIES COMPILE_FLAGS "-msse2")
+        set_source_files_properties(src/mipmap/mipmap_ssse3.c PROPERTIES COMPILE_FLAGS "-mssse3")
+        set_source_files_properties(src/mipmap/mipmap_avx2.c PROPERTIES COMPILE_FLAGS "-mavx2")
+      ELSE ()
+#        neon version not yet tested or implemented
+#        set_source_files_properties(src/mipmap/mipmap_neon.c PROPERTIES COMPILE_FLAGS "-mfpu=neon")
+      ENDIF ()
+  ELSE (NOT MSVC)
+        # try to use sse on x86 based systems
+      IF ( ARCH MATCHES "i386" OR ARCH MATCHES "amd64" OR ARCH MATCHES "x86_64")
+        set_source_files_properties(src/mipmap/mipmap_sse.c PROPERTIES COMPILE_FLAGS "/arch:SSE")
+        set_source_files_properties(src/mipmap/mipmap_sse2.c PROPERTIES COMPILE_FLAGS "/arch:SSE2")
+        set_source_files_properties(src/mipmap/mipmap_ssse3.c PROPERTIES COMPILE_FLAGS "/arch:SSSE3")
       ENDIF ()
   ENDIF (NOT MSVC)
 
@@ -1379,6 +1407,7 @@ ADD_LIBRARY(gorp ${HDRS} ${SRCS}
         ${SRC_S57ENC}
 #        ${SRC_GARMINHOST}
         ${SRC_TEXCMP}
+        ${SRC_MIPMAP}
         ${SRC_SYMBOLS}
         )
