+++ src/glChartCanvas.cpp
@@ -65,7 +65,7 @@
 #include "tcmgr.h"
 #include "compass.h"
 #include "FontMgr.h"
-#include "texcmp/mipmap.h"
+#include "mipmap/mipmap.h"
 
 #ifndef GL_ETC1_RGB8_OES
 #define GL_ETC1_RGB8_OES                                        0x8D64
